src/requirements.txt
pandas
scikit-learn
matplotlib
seaborn
nltk
tensorflow
keras
scikeras
dill
newspaper3k
lime